New Instance 2015

+++ util/jenkins/ansible-provision.sh
@@ -17,9 +17,21 @@
 # - dns_name
 # - environment
 # - name_tag
+set -x
 env
 export PYTHONUNBUFFERED=1
 export BOTO_CONFIG=/var/lib/jenkins/${aws_account}.boto
+
+run_ansible() {
+  ansible-playbook $@
+  ret=$?
+  if [[ $ret -ne 0 ]]; then
+    exit $ret
+  fi
+}
+
+# This DATE_TIME will be used as instance launch time tag
+TERMINATION_DATE_TIME=`date +"%m-%d-%Y %T" --date="-7 days ago"`
 
 if [[ -z $BUILD_USER ]]; then
     BUILD_USER=jenkins
@@ -51,12 +63,13 @@ if [[ ! -f $BOTO_CONFIG ]]; then
 fi
 
 extra_vars_file="/var/tmp/extra-vars-$$.yml"
+sandbox_vars_file="${WORKSPACE}/configuration-secure/ansible/vars/developer-sandbox.yml"
 extra_var_arg="-e@${extra_vars_file}"
 
 if [[ $edx_internal == "true" ]]; then
     # if this is a an edx server include
     # the secret var file
-    extra_var_arg="-e@${extra_vars_file} -e@${WORKSPACE}/configuration-secure/ansible/vars/developer-sandbox.yml"
+    extra_var_arg="-e@${sandbox_vars_file} -e@${extra_vars_file}"
 fi
 
 if [[ -z $region ]]; then
@@ -76,7 +89,7 @@ if [[ -z $elb ]]; then
 fi
 
 if [[ -z $dns_name ]]; then
-  dns_name=$github_username
+  dns_name=${github_username}
 fi
 
 if [[ -z $name_tag ]]; then
@@ -85,11 +98,11 @@ fi
 
 if [[ -z $ami ]]; then
   if [[ $server_type == "full_edx_installation" ]]; then
-    ami="ami-f287419a"
+    ami="ami-b1d92ada"
   elif [[ $server_type == "ubuntu_12.04" || $server_type == "full_edx_installation_from_scratch" ]]; then
-    ami="ami-f478849c"
+    ami="ami-b92bdfd2"
   elif [[ $server_type == "ubuntu_14.04(experimental)" ]]; then
-    ami="ami-a6926dce"
+    ami="ami-3b6a8050"
   fi
 fi
 
@@ -101,25 +114,31 @@ if [[ -z $enable_monitoring ]]; then
   enable_monitoring="false"
 fi
 
+# Lowercase the dns name to deal with an ansible bug
+dns_name="${dns_name,,}"
+
 deploy_host="${dns_name}.${dns_zone}"
 ssh-keygen -f "/var/lib/jenkins/.ssh/known_hosts" -R "$deploy_host"
 
 cd playbooks/edx-east
 
 cat << EOF > $extra_vars_file
----
 ansible_ssh_private_key_file: /var/lib/jenkins/${keypair}.pem
 edx_platform_version: $edxapp_version
 forum_version: $forum_version
+notifier_version: $notifier_version
 xqueue_version: $xqueue_version
 xserver_version: $xserver_version
 ora_version: $ora_version
 ease_version: $ease_version
 certs_version: $certs_version
 discern_version: $discern_version
+configuration_version: $configuration_version
 EDXAPP_STATIC_URL_BASE: $static_url_base
 EDXAPP_LMS_NGINX_PORT: 80
 EDXAPP_LMS_PREVIEW_NGINX_PORT: 80
+EDXAPP_CMS_NGINX_PORT: 80
+NGINX_SET_X_FORWARDED_HEADERS: True
 EDX_ANSIBLE_DUMP_VARS: true
 migrate_db: "yes"
 openid_workaround: True
@@ -166,13 +185,15 @@ COMMON_USER_INFO:
     github: true
     type: admin
 USER_CMD_PROMPT: '[$name_tag] '
-COMMON_ENABLE_NEWRELIC: $enable_monitoring
+COMMON_ENABLE_NEWRELIC_APP: $enable_monitoring
 COMMON_ENABLE_DATADOG: $enable_monitoring
 FORUM_NEW_RELIC_ENABLE: $enable_monitoring
 EDXAPP_NEWRELIC_LMS_APPNAME: sandbox-${dns_name}-edxapp-lms
 EDXAPP_NEWRELIC_CMS_APPNAME: sandbox-${dns_name}-edxapp-cms
+EDXAPP_NEWRELIC_WORKERS_APPNAME: sandbox-${dns_name}-edxapp-workers
 XQUEUE_NEWRELIC_APPNAME: sandbox-${dns_name}-xqueue
 FORUM_NEW_RELIC_APP_NAME: sandbox-${dns_name}-forums
+SANDBOX_USERNAME: $github_username
 EOF
 fi
 
@@ -193,6 +214,7 @@ instance_tags:
     Name: $name_tag
     source: jenkins
     owner: $BUILD_USER
+    instance_termination_time: $TERMINATION_DATE_TIME
     datadog: monitored
 root_ebs_size: $root_ebs_size
 name_tag: $name_tag
@@ -205,18 +227,18 @@ EOF
 
     # run the tasks to launch an ec2 instance from AMI
     cat $extra_vars_file
-    ansible-playbook edx_provision.yml -i inventory.ini $extra_var_arg --user ubuntu
+    run_ansible edx_provision.yml -i inventory.ini $extra_var_arg --user ubuntu
 
     if [[ $server_type == "full_edx_installation" ]]; then
         # additional tasks that need to be run if the
         # entire edx stack is brought up from an AMI
-        ansible-playbook rabbitmq.yml -i "${deploy_host}," $extra_var_arg --user ubuntu
-        ansible-playbook restart_supervisor.yml -i "${deploy_host}," $extra_var_arg --user ubuntu
+        run_ansible rabbitmq.yml -i "${deploy_host}," $extra_var_arg --user ubuntu
+        run_ansible restart_supervisor.yml -i "${deploy_host}," $extra_var_arg --user ubuntu
     fi
 fi
 
 declare -A deploy
-roles="edxapp forum xqueue xserver ora discern certs demo"
+roles="edxapp forum notifier xqueue xserver ora discern certs demo testcourses"
 for role in $roles; do
     deploy[$role]=${!role}
 done
@@ -225,7 +247,7 @@ done
 # run non-deploy tasks for all roles
 if [[ $reconfigure == "true" || $server_type == "full_edx_installation_from_scratch" ]]; then
     cat $extra_vars_file
-    ansible-playbook edx_continuous_integration.yml -i "${deploy_host}," $extra_var_arg --user ubuntu
+    run_ansible edx_continuous_integration.yml -i "${deploy_host}," $extra_var_arg --user ubuntu
 fi
 
 if [[ $reconfigure != "true" && $server_type == "full_edx_installation" ]]; then
@@ -233,15 +255,22 @@ if [[ $reconfigure != "true" && $server_type == "full_edx_installation" ]]; then
     for i in $roles; do
         if [[ ${deploy[$i]} == "true" ]]; then
             cat $extra_vars_file
-            ansible-playbook ${i}.yml -i "${deploy_host}," $extra_var_arg --user ubuntu --tags deploy
+            run_ansible ${i}.yml -i "${deploy_host}," $extra_var_arg --user ubuntu --tags deploy
         fi
     done
 fi
 
 # deploy the edx_ansible role
-ansible-playbook edx_ansible.yml -i "${deploy_host}," $extra_var_arg --user ubuntu
+run_ansible edx_ansible.yml -i "${deploy_host}," $extra_var_arg --user ubuntu
+cat $sandbox_vars_file $extra_vars_file | grep -v -E "_version|migrate_db" > ${extra_vars_file}_clean
+ansible -c ssh -i "${deploy_host}," $deploy_host -m copy -a "src=${extra_vars_file}_clean dest=/edx/app/edx_ansible/server-vars.yml" -u ubuntu -b
+ret=$?
+if [[ $ret -ne 0 ]]; then
+  exit $ret
+fi
 
 # set the hostname
-ansible-playbook set_hostname.yml -i "${deploy_host}," -e hostname_fqdn=${deploy_host} --user ubuntu
+run_ansible set_hostname.yml -i "${deploy_host}," -e hostname_fqdn=${deploy_host} --user ubuntu
 
 rm -f "$extra_vars_file"
+rm -f ${extra_vars_file}_clean
