+++ util/jenkins/check-migrations.sh
@@ -1,11 +1,8 @@
 #!/usr/bin/env bash
 set -x
+set -e
 
-if [[
-        -z $WORKSPACE       ||
-        -z $environment     ||
-        -z $deployment
-    ]]; then
+if [[ -z $WORKSPACE ]]; then
     echo "Environment incorrect for this wrapper script"
     env
     exit 1
@@ -13,7 +10,7 @@ fi
 
 
 env
-cd $WORKSPACE/edx-platform
+cd "$WORKSPACE/edx-platform"
 
 # install requirements
 # These requirements will be installed into the shinginpanda
@@ -22,28 +19,56 @@ cd $WORKSPACE/edx-platform
 
 pip install --exists-action w -r requirements/edx/pre.txt
 pip install --exists-action w -r requirements/edx/base.txt
-pip install --exists-action w -r requirements/edx/post.txt
-pip install --exists-action w -r requirements/edx/repo.txt
+if [[ -f requirements/edx/post.txt ]]; then
+  pip install --exists-action w -r requirements/edx/post.txt
+fi
+if [[ -f requirements/edx/repo.txt ]]; then
+  pip install --exists-action w -r requirements/edx/repo.txt
+fi
 pip install --exists-action w -r requirements/edx/github.txt
-pip install --exists-action w -r requirements/edx/local.txt
+if [[ -f requirements/edx/local.txt ]]; then
+  pip install --exists-action w -r requirements/edx/local.txt
+fi
+pip install --exists-action w -r requirements/edx/edx-private.txt
 
-cd $WORKSPACE/configuration/playbooks/edx-east
+if [[ $openid_workaround == "true" ]]; then
+    sed -i -e 's/claimed_id = models.TextField(max_length=2047, unique=True/claimed_id = models.TextField(max_length=2047/'  "$VIRTUAL_ENV/lib/python2.7/site-packages/django_openid_auth/models.py"
+fi
+
+cd "$WORKSPACE/configuration/playbooks/edx-east"
 
 if [[ -f ${WORKSPACE}/configuration-secure/ansible/vars/${deployment}.yml ]]; then
     extra_var_args+=" -e@${WORKSPACE}/configuration-secure/ansible/vars/${deployment}.yml"
 fi
 
-if [[ $db_dry_run=="false" ]]; then
-    # Set this to an empty string if db_dry_run is 
-    # not set.  By default the db_dry_run var is
-    # set to --db-dry-run
-    extra_var_args+=" -e db_dry_run=''"
+if [[ -z $syncdb ]]; then
+    syncdb="false"
 fi
 
-extra_var_args+=" -e@${WORKSPACE}/configuration-secure/ansible/vars/${environment}-${deployment}.yml"
+if [[ $db_dry_run == "false" ]]; then
+    # Set this to an empty string if db_dry_run is
+    # not set.  By default the db_dry_run var is
+    # set to --db-dry-run
+
+    extra_var_args+=" -e db_dry_run=''"
+else
+    # always skip syncdb unless dry run is unchecked
+    syncdb="false"
+fi
+
+if [[ -f ${WORKSPACE}/configuration-secure/ansible/vars/${environment}-${deployment}.yml ]]; then
+    extra_var_args+=" -e@${WORKSPACE}/configuration-secure/ansible/vars/${environment}-${deployment}.yml"
+fi
+
+for extra_var in $extra_vars; do
+    extra_var_args+=" -e@${WORKSPACE}/configuration-secure/ansible/vars/$extra_var"
+done
+
 extra_var_args+=" -e edxapp_app_dir=${WORKSPACE}"
 extra_var_args+=" -e edxapp_code_dir=${WORKSPACE}/edx-platform"
 extra_var_args+=" -e edxapp_user=jenkins"
+extra_var_args+=" -e syncdb=$syncdb"
+extra_var_args+=" -e EDXAPP_CFG_DIR=${WORKSPACE}"
 
 # Generate the json configuration files
 ansible-playbook -c local $extra_var_args --tags edxapp_cfg -i localhost, -s -U jenkins edxapp.yml
